+++ source/STAR.cpp
@@ -416,7 +416,13 @@ int main(int argInN, char* argIn[]) {
     *P->inOut->logStdOut << timeMonthDayTime(g_statsAll.timeFinish) << " ..... Finished successfully\n" <<flush;
 
     P->inOut->logMain  << "ALL DONE!\n"<<flush;
-    sysRemoveDir (P->outFileTmp);
+    if (P->outTmpKeep=="None")
+    {
+        sysRemoveDir (P->outFileTmp);
+    } else
+    {
+        //nothing to do
+    };
 
     P->closeReadsFiles();//this will kill the readFilesCommand processes if necessary
     mainGenome.~Genome(); //need explicit call because of the 'delete P->inOut' below, which will destroy P->inOut->logStdOut